+++ tsMuxer/mlpCodec.h
@@ -26,11 +26,11 @@ class MLPCodec
     uint64_t getFrameDuration() const;
     static int mlp_samplerate(int ratebits);
 
-    int m_channels;
+    uint8_t m_channels;
     int m_samples;
     int m_samplerate;  // Sample rate of first substream
     int m_bitrate;     // Peak bitrate for VBR, actual bitrate for CBR
-    int m_substreams;
+    uint8_t m_substreams;
     MlpSubType m_subType;
 };
