+++ tsMuxer/vvc.h
@@ -58,8 +58,8 @@ struct VvcUnit
     int nalBufferLen() const { return m_nalBufferLen; }
 
     NalType nal_unit_type;
-    int nuh_layer_id;
-    int nuh_temporal_id_plus1;
+    uint8_t nuh_layer_id;
+    uint8_t nuh_temporal_id_plus1;
 
    protected:
     unsigned extractUEGolombCode();
@@ -80,12 +80,12 @@ struct VvcUnitWithProfile : VvcUnit
     VvcUnitWithProfile();
     std::string getProfileString() const;
 
-    int profile_idc;
-    int tier_flag;
-    int level_idc;
+    uint8_t profile_idc;
+    uint8_t tier_flag;
+    uint8_t level_idc;
     bool ptl_frame_only_constraint_flag;
-    int ptl_num_sub_profiles;
-    std::vector<int> general_sub_profile_idc;
+    uint8_t ptl_num_sub_profiles;
+    std::vector<unsigned> general_sub_profile_idc;
 
    protected:
     int profile_tier_level(bool profileTierPresentFlag, int MaxNumSubLayersMinus1);
@@ -99,9 +99,9 @@ struct VvcVpsUnit : VvcUnitWithProfile
     void setFPS(double fps);
     std::string getDescription() const;
 
-    int vps_id;
-    int vps_max_layers;
-    int vps_max_sublayers;
+    uint8_t vps_id;
+    uint8_t vps_max_layers;
+    uint8_t vps_max_sublayers;
     int num_units_in_tick;
     int time_scale;
     int num_units_in_tick_bit_pos;
@@ -115,14 +115,14 @@ struct VvcSpsUnit : VvcUnitWithProfile
     double getFPS() const;
     std::string getDescription() const;
 
-    int sps_id;
-    int vps_id;
-    int max_sublayers_minus1;
-    int chroma_format_idc;
+    uint8_t sps_id;
+    uint8_t vps_id;
+    uint8_t max_sublayers_minus1;
+    uint8_t chroma_format_idc;
     unsigned pic_width_max_in_luma_samples;
     unsigned pic_height_max_in_luma_samples;
     unsigned bitdepth_minus8;
-    int log2_max_pic_order_cnt_lsb;
+    uint8_t log2_max_pic_order_cnt_lsb;
     VvcHrdUnit m_sps_hrd;
 
     std::vector<unsigned> cpb_cnt_minus1;
@@ -131,9 +131,9 @@ struct VvcSpsUnit : VvcUnitWithProfile
     bool progressive_source_flag;
     bool interlaced_source_flag;
     bool non_packed_constraint_flag;
-    int colour_primaries;
-    int transfer_characteristics;
-    int matrix_coeffs;
+    uint8_t colour_primaries;
+    uint8_t transfer_characteristics;
+    uint8_t matrix_coeffs;
     bool full_range_flag;
 
    private:
@@ -151,8 +151,8 @@ struct VvcPpsUnit : VvcUnit
     VvcPpsUnit();
     int deserialize();
 
-    int pps_id;
-    int sps_id;
+    uint8_t pps_id;
+    uint8_t sps_id;
 };
 
 struct VvcSliceHeader : VvcUnit
@@ -161,9 +161,9 @@ struct VvcSliceHeader : VvcUnit
     int deserialize(const VvcSpsUnit* sps, const VvcPpsUnit* pps);
     bool isIDR() const;
 
-    int pic_order_cnt_lsb;
+    uint16_t pic_order_cnt_lsb;
 };
 
-std::vector<std::vector<uint8_t>> vvc_extract_priv_data(const uint8_t* buff, int size, int* nal_size);
+std::vector<std::vector<uint8_t>> vvc_extract_priv_data(const uint8_t* buff, int size, uint8_t* nal_size);
 
 #endif  // VVC_H_
